requirements.txt
absl-py==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.16
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
av==12.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
einops==0.8.2
exceptiongroup==1.3.1
face-alignment==1.4.1
fastapi==0.115.6
ffmpy==1.0.0
filelock==3.20.0
flow-vis==0.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gradio==5.1.0
gradio_client==1.4.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.15.2
llvmlite==0.46.0
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.4.2
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
orjson==3.11.7
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.1.4
pillow==10.4.0
platformdirs==4.5.1
pooch==1.9.0
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytorch-lightning==2.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
ruff==0.15.0
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.41.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.9
tokenizers==0.13.3
tomlkit==0.12.0
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchdiffeq==0.2.5
torchmetrics==1.8.2
torchvision==0.18.1+cu121
tqdm==4.67.3
transformers==4.30.2
triton==2.3.1
typer==0.21.1
typing_extensions==4.15.0
tzdata==2025.3
unzip==1.0.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==12.0
Werkzeug==3.1.5
yarl==1.22.0
